+++ src/penny_ante/__about__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: 2024-present Peter Bowen <peter@bowenfamily.org>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "0.1.0"
+__version__ = "1.0.0"